+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to this project will be documented in this file.
 
+## Master
+
+Adds Backpack text styles. See th full list in the [README.md](README.md#text-styles).
+
+Usage:
+
+```XML
+<TextView
+  style="@style/bpkTextSm" />
+```
+
 ## 0.0.3
 
 Update project id to "net.skyscanner.backpack"